general adjust for the new schema in progress
