+++ src/@types/veterinaryProps.d.ts
@@ -1,4 +1,15 @@
 type createVeterinaryController = {
+	personName: string;
+	cpf: string;
+	email: string;
+	password: string;
+	cellphoneNumber: string;
+	phoneNumber: string | null;
+	address: {
+		zipCode: string;
+		number: string;
+		complement: string | null;
+	};
 	occupationArea: string;
 	formation: string;
 	institution: string;
