+++ src/model/petModel.ts
@@ -1,5 +1,5 @@
-import { PetGender, PetSize, PetSpecie } from '@prisma/client';
 import prisma from '../lib/prisma';
+import {CreatePetInfosModelProps, UpdatePetInfosModelProps} from "../@types/petInfosProps";
 
 export default class Pet {
 	async findPet(petID: number) {
@@ -9,8 +9,6 @@ export default class Pet {
 					id: petID,
 				},
 				include: {
-					petSize: true,
-					petGender: true,
 					petSpecie: true,
 				},
 			});
@@ -39,20 +37,18 @@ export default class Pet {
 	}
 	async createNewPet(pet: CreatePetInfosModelProps) {
 		try {
-			const createdPet = await prisma.pet.create({
+			return await prisma.pet.create({
 				data: {
 					name: pet.name,
 					birthDate: pet.birthDate,
 					microship: pet.microship,
-					// @ts-ignore
 					ownerId: pet.ownerId,
-					petGenderId: pet.genderId,
-					petSizeId: pet.sizeId,
+					petGender: pet.gender,
+					petSize: pet.size,
 					petSpecieId: pet.specieId,
 					photo: pet.photo,
 				},
 			});
-			return createdPet;
 		} catch (err) {
 			console.log(err);
 			throw new Error(`${err}`);
@@ -65,8 +61,6 @@ export default class Pet {
 					id: petID,
 				},
 				include: {
-					petSize: true,
-					petGender: true,
 					petSpecie: true,
 				},
 			});
@@ -85,8 +79,8 @@ export default class Pet {
 					name: petInfos.name,
 					birthDate: petInfos.birthDate,
 					microship: petInfos.microship,
-					petGenderId: petInfos.genderId,
-					petSizeId: petInfos.sizeId,
+					petGender: petInfos.gender,
+					petSize: petInfos.size,
 					petSpecieId: petInfos.specieId,
 					photo: petInfos.photo,
 				},
@@ -97,35 +91,5 @@ export default class Pet {
 			console.log(err);
 			throw new Error(`${err}`);
 		}
-	}
-}
-
-export class PetComplements {
-	async getSpecie(specie: string): Promise<PetSpecie | null> {
-		const foundSpecie = await prisma.petSpecie.findMany({
-			where: {
-				name: specie,
-			},
-		});
-		if (foundSpecie.length > 0) return foundSpecie[0];
-		return null;
-	}
-	async getGender(gender: string): Promise<PetGender | null> {
-		const foundGender = await prisma.petGender.findMany({
-			where: {
-				name: gender,
-			},
-		});
-		if (foundGender.length > 0) return foundGender[0];
-		return null;
-	}
-	async getSize(size: string): Promise<PetSize | null> {
-		const foundSize = await prisma.petSize.findMany({
-			where: {
-				name: size,
-			},
-		});
-		if (foundSize.length > 0) return foundSize[0];
-		return null;
 	}
 }
